# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chat audio API Documentation",
        "description": "API documentation for the Chat audio application",
        "contact": {
            "email": "tsiresymila@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/chats/{chat}/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Get chat messages",
                "description": "Retrieve all messages for a specific chat",
                "operationId": "ed8820ee981e0aa106f48ac69983e460",
                "parameters": [
                    {
                        "name": "chat",
                        "in": "path",
                        "description": "Chat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messages retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "content": {
                                                        "type": "string"
                                                    },
                                                    "type": {
                                                        "type": "string"
                                                    },
                                                    "user_id": {
                                                        "type": "integer"
                                                    },
                                                    "chat_id": {
                                                        "type": "integer"
                                                    },
                                                    "is_read": {
                                                        "type": "boolean"
                                                    },
                                                    "audio_path": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Chat not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Create a new message",
                "description": "Create a new message in a chat (text or audio)",
                "operationId": "dfdf2c894530860f4aa17a26f040881e",
                "parameters": [
                    {
                        "name": "chat",
                        "in": "path",
                        "description": "Chat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "content",
                                    "type"
                                ],
                                "properties": {
                                    "content": {
                                        "description": "Message content or transcription",
                                        "type": "string"
                                    },
                                    "type": {
                                        "description": "Message type",
                                        "type": "string",
                                        "enum": [
                                            "text",
                                            "audio"
                                        ]
                                    },
                                    "audio": {
                                        "description": "Audio file (required if type is audio)",
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "stream": {
                                        "description": "Whether to stream the AI response",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "content": {
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "chat_id": {
                                                    "type": "integer"
                                                },
                                                "is_read": {
                                                    "type": "boolean"
                                                },
                                                "audio_path": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "ai_response": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "content": {
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "chat_id": {
                                                    "type": "integer"
                                                },
                                                "is_read": {
                                                    "type": "boolean"
                                                },
                                                "audio_path": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Chat not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chats/{chat}/messages/transcribe": {
            "post": {
                "summary": "Transcribe audio to text",
                "description": "Transcribes an uploaded audio file and returns the transcription.",
                "operationId": "4bb05c7dfcbb933deab6b2ebda6bce7a",
                "parameters": [
                    {
                        "name": "chat",
                        "in": "path",
                        "description": "Chat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "audio"
                                ],
                                "properties": {
                                    "audio": {
                                        "description": "Audio file",
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transcription": {
                                            "type": "string"
                                        },
                                        "language": {
                                            "type": "string"
                                        },
                                        "audio_path": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Chat not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}